+++ plugins/rmanipulation/taskmanipulation.cpp
@@ -658,17 +658,18 @@ protected:
                 Transform t = _robot->GetTransform();
                 _phandtraj->GetConfigurationSpecification().ExtractTransform(t,vtrajdata.begin(),_robot);
 
+                Vector vglobalpalmdir;
+                if( iGraspDir >= 0 ) {
+                    vglobalpalmdir = transTarg.rotate(Vector(pgrasp[iGraspDir], pgrasp[iGraspDir+1], pgrasp[iGraspDir+2]));
+                }
+                else {
+                    vglobalpalmdir = pmanip->GetTransform().rotate(pmanip->GetDirection());
+                }
+
                 // move back a little if robot/target in collision
                 if( !!ptarget ) {
                     RobotBase::RobotStateSaver saverlocal(_robot);
                     _robot->SetTransform(t);
-                    Vector vglobalpalmdir;
-                    if( iGraspDir >= 0 ) {
-                        vglobalpalmdir = transTarg.rotate(Vector(pgrasp[iGraspDir], pgrasp[iGraspDir+1], pgrasp[iGraspDir+2]));
-                    }
-                    else {
-                        vglobalpalmdir = pmanip->GetTransform().rotate(pmanip->GetDirection());
-                    }
                     dReal fstep=0;
                     dReal fstepbacksize = 0.001f;
                     while(GetEnv()->CheckCollision(KinBodyConstPtr(_robot),KinBodyConstPtr(ptarget))) {
@@ -679,14 +680,28 @@ protected:
                     }
                     if( fstep > 0 ) {
                         RAVELOG_DEBUG(str(boost::format("grasp %d: moved %f along direction=[%f,%f,%f]")%igrasp%fstep% -vglobalpalmdir.x% -vglobalpalmdir.y% -vglobalpalmdir.z));
-                        if( fGraspApproachOffset < 0 ) {
-                            RAVELOG_WARN(str(boost::format("grasp %d: moved too far back to avoid collision, approach offset is now negative (%f) and cannot recover. Should increase approachoffset")%igrasp%fGraspApproachOffset));
-                        }
                     }
                 }
 
                 // find the end effector transform
                 tGoalEndEffector.SetTransform6D(t * _robot->GetTransform().inverse() * pmanip->GetTransform());
+
+                if( iGraspTransform >= 0 ) {
+                    // use the grasp transform to figure out how much backing to compensate for, this is just a sanity check
+                    dReal* pm = pgrasp+iGraspTransform;
+                    TransformMatrix tmexpected;
+                    tmexpected.m[0] = pm[0]; tmexpected.m[1] = pm[3]; tmexpected.m[2] = pm[6]; tmexpected.trans.x = pm[9];
+                    tmexpected.m[4] = pm[1]; tmexpected.m[5] = pm[4]; tmexpected.m[6] = pm[7]; tmexpected.trans.y = pm[10];
+                    tmexpected.m[8] = pm[2]; tmexpected.m[9] = pm[5]; tmexpected.m[10] = pm[8]; tmexpected.trans.z = pm[11];
+                    Transform texpectedglobal = ptarget->GetTransform() * Transform(tmexpected);
+                    dReal dist = vglobalpalmdir.dot3(tGoalEndEffector.GetTransform6D().trans-texpectedglobal.trans);
+                    fGraspApproachOffset = fApproachOffset+dist;
+                }
+
+                if( fGraspApproachOffset < 0 ) {
+                    RAVELOG_WARN(str(boost::format("grasp %d: moved too far back to avoid collision, approach offset is now negative (%f) and cannot recover. Should increase approachoffset")%igrasp%fGraspApproachOffset));
+                }
+
                 vFinalGripperValues.resize(pmanip->GetGripperIndices().size(),0);
                 _phandtraj->GetConfigurationSpecification().ExtractJointValues(vFinalGripperValues.begin(),vtrajdata.begin(),_robot,pmanip->GetGripperIndices());
             }
@@ -786,10 +801,6 @@ protected:
                         ss << "]" << endl;
                         RAVELOG_DEBUG(ss.str());
                     }
-                }
-                else {
-                    // revert back?
-                    //tApproachEndEffector.trans += fSmallOffset * vglobalpalmdir;
                 }
             }
 
@@ -964,6 +975,7 @@ protected:
         }
 
         if( bCombinePreShapeTraj ) {     // add the preshape
+            RAVELOG_DEBUG(str(boost::format("combine preshape trajectory, duration=%f")%itpreshapetraj->second->GetDuration()));
             itpreshapetraj->second->GetWaypoints(0,itpreshapetraj->second->GetNumWaypoints(),vtrajdata,specfinal);
             ptrajfinal->Insert(ptrajfinal->GetNumWaypoints(),vtrajdata);
             // set the last point so the converters can pick it up